Rename rpi and hfi readers
